Revert the parent-pid changes in order to land a refactoring pr instead.
